Extract tool-surface resolution into selection module
Move the model-callable tool resolution logic out of agent.rs into
tool_surface::selection, including the resolve_with_subagent_tools_for_mcp_presence
refactor that takes mcp_services_online as a plain bool. Pure, behavior-preserving
extraction so the same resolution can be reused outside the async agent path.
