+++ crates/defra-agent/src/tool_surface/selection.rs
@@ -8,7 +8,10 @@ use super::modes::{BashMode, FileToolMode};
 use std::path::PathBuf;
 
 use crate::document_config::SubagentTarget;
-use crate::toolset::CommandExecutionPolicy;
+use crate::toolset::{
+    default_read_only_command_policy, parse_argv_prefixes, CommandExecutionMode,
+    CommandExecutionPolicy, CommandNetworkMode,
+};
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub(crate) struct SubagentToolConfig {
@@ -24,6 +27,35 @@ pub(crate) struct SubagentToolConfig {
 }
 
 impl SubagentToolConfig {
+    pub(crate) fn from_document(selection: &crate::document_config::ToolSelectionDocument) -> Self {
+        let targets = selection
+            .subagent_targets
+            .iter()
+            .flatten()
+            .filter_map(
+                |entry| match crate::document_config::SubagentTarget::parse(entry) {
+                    Ok(target) => Some(target),
+                    Err(error) => {
+                        tracing::warn!(
+                            selection_id = %selection.selection_id,
+                            entry = %entry,
+                            %error,
+                            "skipping malformed subagent_targets entry"
+                        );
+                        None
+                    }
+                },
+            )
+            .collect();
+        Self {
+            targets,
+            spawn_enabled: selection.subagent_spawn_enabled.unwrap_or(false),
+            steering_enabled: selection.subagent_steering_enabled.unwrap_or(false),
+            background_enabled: selection.subagent_background_enabled.unwrap_or(false),
+            allow_cross_deployment: selection.subagent_allow_cross_deployment.unwrap_or(false),
+        }
+    }
+
     pub(crate) fn tools_enabled(&self) -> bool {
         self.spawn_enabled && !self.targets.is_empty()
     }
@@ -85,6 +117,131 @@ impl Default for ToolSelection {
             defra_query_collections: Vec::new(),
         }
     }
+}
+
+impl ToolSelection {
+    pub(crate) fn from_document(
+        selection: &crate::document_config::ToolSelectionDocument,
+    ) -> anyhow::Result<Self> {
+        let bash = if selection.enable_bash.unwrap_or(false) {
+            BashMode::parse(selection.bash_mode.as_deref().unwrap_or("ReadOnly"))?
+        } else {
+            BashMode::Off
+        };
+        Ok(Self {
+            file_tools: if selection.enable_file_tools.unwrap_or(false) {
+                FileToolMode::parse(selection.file_tools_mode.as_deref().unwrap_or("ReadOnly"))?
+            } else {
+                FileToolMode::Off
+            },
+            file_tool_root: selection
+                .file_tool_root
+                .as_deref()
+                .and_then(normalize_optional_string)
+                .map(PathBuf::from),
+            bash,
+            command_policy: command_policy_from_document(selection, bash)?,
+            cli_tool_names: selection.cli_tool_names.clone().unwrap_or_default(),
+            enable_meta_tools: selection.enable_meta_tools.unwrap_or(true),
+            allowed_mcp_service_ids: selection
+                .allowed_mcp_service_ids
+                .clone()
+                .unwrap_or_default(),
+            backgroundable_tool_names: selection
+                .backgroundable_tool_names
+                .clone()
+                .unwrap_or_default(),
+            enable_memory: selection.enable_memory.unwrap_or(false),
+            enable_session_history_tool: selection.enable_session_history_tool.unwrap_or(false),
+            enable_defra_query: selection.enable_defra_query.unwrap_or(true),
+            defra_query_collections: selection
+                .defra_query_collections
+                .clone()
+                .unwrap_or_default(),
+        })
+    }
+}
+
+fn command_policy_from_document(
+    selection: &crate::document_config::ToolSelectionDocument,
+    bash: BashMode,
+) -> anyhow::Result<Option<CommandExecutionPolicy>> {
+    let has_policy = selection
+        .command_execution_policy
+        .as_deref()
+        .and_then(normalize_optional_string)
+        .is_some()
+        || selection
+            .command_network_mode
+            .as_deref()
+            .and_then(normalize_optional_string)
+            .is_some()
+        || selection
+            .command_allowed_argv_prefixes
+            .as_ref()
+            .is_some_and(|prefixes| !prefixes.is_empty())
+        || selection
+            .command_forbidden_argv_prefixes
+            .as_ref()
+            .is_some_and(|prefixes| !prefixes.is_empty());
+    if !has_policy {
+        return if matches!(bash, BashMode::Unrestricted) {
+            Ok(Some(
+                CommandExecutionPolicy::write_capable()
+                    .with_mode(CommandExecutionMode::Unrestricted),
+            ))
+        } else {
+            Ok(None)
+        };
+    }
+
+    let requested_mode = selection
+        .command_execution_policy
+        .as_deref()
+        .and_then(normalize_optional_string)
+        .map(CommandExecutionMode::parse)
+        .transpose()?;
+    let mode = match bash {
+        BashMode::Off | BashMode::ReadOnly => CommandExecutionMode::ReadOnly,
+        BashMode::Unrestricted => requested_mode.unwrap_or(CommandExecutionMode::Unrestricted),
+    };
+
+    let allowed = parse_argv_prefixes(
+        selection
+            .command_allowed_argv_prefixes
+            .as_deref()
+            .unwrap_or(&[]),
+    )?;
+    let forbidden = parse_argv_prefixes(
+        selection
+            .command_forbidden_argv_prefixes
+            .as_deref()
+            .unwrap_or(&[]),
+    )?;
+    let network_mode = selection
+        .command_network_mode
+        .as_deref()
+        .and_then(normalize_optional_string)
+        .map(CommandNetworkMode::parse)
+        .transpose()?
+        .unwrap_or(CommandNetworkMode::Inherit);
+
+    let base = if matches!(mode, CommandExecutionMode::ReadOnly) {
+        default_read_only_command_policy()
+    } else {
+        CommandExecutionPolicy::write_capable()
+    };
+    Ok(Some(
+        base.with_mode(mode)
+            .with_allowed_argv_prefixes(allowed)
+            .with_forbidden_argv_prefixes(forbidden)
+            .with_network_mode(network_mode),
+    ))
+}
+
+fn normalize_optional_string(value: &str) -> Option<&str> {
+    let trimmed = value.trim();
+    (!trimmed.is_empty()).then_some(trimmed)
 }
 
 type CustomToolFactoryFn = Arc<dyn Fn() -> Result<Box<dyn ToolDyn>> + Send + Sync>;
